+++ examples/virtual/src/mod.ts
@@ -1,0 +1,3 @@
+import logMessage from "log";
+
+logMessage();