+++ plugins/virtual/mod.ts
@@ -1,0 +1,1 @@
+export { virtual } from "./deps.ts";